Load sphinx config in a separate process
This loads the Sphinx config in a different process to deal with
module-reload issues. These can occur when the `conf.py` file imports a
module from the local git repo to set the version and release of the
current docs.

Before, all `conf.py` files were read in the same interpreter process,
so the module would only be loaded once. This is undesired if the module
is different for each version to build docs for. For example, changes in
the `__version__` variable would not be picked up and it might lead to
errors when `conf.py` and module go out of sync.

The alternative to this method would be to iterate over `sys.modules`
and reset/reload them manually, but that would probably require
additional configuration by the user and looks even hackier than this
implementation.