Add ability to configure multiple build targets
Add smv_build_targets configuration option. This will allow users to
specify an arbitrary number of build targets to build for each version
of the documentation, and specify whether it should be available to
documentation consumers to download. This supports all sphinx builders,
and output formats where there is a single output file. For build
targets where there are multiple output files (such as html), all archive
formats supported by shutil.make_archive are supported. Downloadable
build targets are placed in an `artefacts/` directory within each
versions build directory.

Add smv_clean_intermediate_files configuration option. This will allow
users to determine whether the intermediate files are removed or not.
This is a useful configuration option where the generated build
artefacts are not for public download consumption.

Reorder build arguments passed to `sphinx` such that the source and
build directory are the first two arguments. This is to ensure that
there is compatability between the `-M` and `-b` sphinx build flags,
which have slightly different APIs.