Add more advanced artefact file matching
Add file matching that accounts for project names that include spaces
and is case insensitive. This is to match how latex and other builders
generate their artefacts.

Use only the project name from the build configuration rather than the
configuration on each branch to account for when the project name has
been changed to display differently.