Add support to locate artefacts from nested paths
Trying to access a download artefact from a page other than the root
would fail to resolve the correct artefact path. This change adds
support to resolve the correct path from any page the user is located
on.

Add unit tests to confirm this feature works as expected, and also to
ensure artefacts are found correctly when the project name contains
spaces and variable casing.