ci: Implement a prim and proper release workflow
Always build the GH Pages from the default branch to ensure that the
documentation is always built with the _latest_ SMV and not e.g. a patch
version for an older release line.

Move the actual build pipeline into its own reusable workflow in order
to allow fine grained builds of the documentation and release artifacts.

Add a CI job to create a GH release draft for a new tag which
automatically attaches the released build artifacts.